+++ src/ch/nivisan/rain/entity/projectile/WizardProjectile.java
@@ -7,17 +7,17 @@ import ch.nivisan.rain.level.Level;
 
 public class WizardProjectile extends Projectile {
     // refers to the rate of how many per second (60 = 1)
-    public static final double fireRate = 30;
+    public static final float fireRate = 30;
 
-    public WizardProjectile(int xOrigin, int yOrigin, double angle, Level level) {
+    public WizardProjectile(float xOrigin, float yOrigin, double angle, Level level) {
         super(xOrigin, yOrigin, angle, Sprite.wizardProjectile, level);
         range = 200;
         damage = 20;
         speed = 4;
 
         // angle and vector length
-        nx = speed * Math.cos(angle);
-        ny = speed * Math.sin(angle);
+        nx = (float)(speed * Math.cos(angle));
+        ny = (float)(speed * Math.sin(angle));
     }
 
     public void update() {
@@ -42,9 +42,9 @@ public class WizardProjectile extends Projectile {
      *
      * @return
      */
-    private double calculateDistance() {
-        double distance = 0;
-        distance = Math.sqrt(Math.abs(((xOrigin - x) * (xOrigin - x)) + ((yOrigin - y) * (yOrigin - y))));
+    private float calculateDistance() {
+        float distance = 0;
+        distance = (float) Math.sqrt(Math.abs(((xOrigin - x) * (xOrigin - x)) + ((yOrigin - y) * (yOrigin - y))));
         return distance;
     }
